changed everything over from doubles to floats - perhaps that will fix the artifacts

+++ add_random_colors.c
@@ -34,12 +34,12 @@ int main(int argc, char ** argv)
       va->array_type = GL_COLOR_ARRAY;
       va->num_dimensions = 4;
       va->shape = shape;
-      va->vertexs = (double*)malloc(sizeof(double)*va->num_dimensions*shape->num_vertexs);
+      va->vertexs = (float*)malloc(sizeof(float)*va->num_dimensions*shape->num_vertexs);
       long i;
-      double r = rand() / (float)RAND_MAX / 2.0 + 0.2;
-      double g = rand() / (float)RAND_MAX / 2.0 + 0.2;
-      double b = rand() / (float)RAND_MAX / 2.0 + 0.5;
-      double a = rand() / (float)RAND_MAX / 4.0 + 0.5;
+      float r = rand() / (float)RAND_MAX / 2.0 + 0.2;
+      float g = rand() / (float)RAND_MAX / 2.0 + 0.2;
+      float b = rand() / (float)RAND_MAX / 2.0 + 0.5;
+      float a = rand() / (float)RAND_MAX / 4.0 + 0.5;
       for (i = 0 ; i < shape->num_vertexs ; i++)
       {
         va->vertexs[i*va->num_dimensions+0] = r;
